ml-service-flask/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
pymupdf==1.23.14
pdfplumber==0.10.3
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
openai==1.6.1
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0
langdetect==1.0.9
textblob==0.17.1
yake==0.4.8
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2
joblib==1.3.2
